requirements.txt
boto3==1.34.0
firebase_admin==6.3.0
folium==0.16.0
joblib==1.2.0
langchain_google_genai==1.0.4
matplotlib==3.7.1
opencv_contrib_python==4.7.0.72
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.76
pandas==1.5.3
pandasai==2.0.42
plotly==5.18.0
protobuf<5
python-dotenv==1.0.1
Requests==2.31.0
scikit_learn==1.2.2
streamlit==1.34.0
streamlit_elements==0.1.0
streamlit_folium==0.19.0
streamlit_option_menu==0.3.12
streamlit_shadcn_ui==0.1.18
streamlit_survey==0.1.0
